+++ src/libraries/Address.sol
@@ -9,9 +9,7 @@ library Address {
     bytes16 private constant _CAPITAL = "0123456789ABCDEF";
 
     ///@dev convert address to checksum address string.
-    function toChecksumString(
-        address addr
-    ) internal pure returns (string memory) {
+    function toChecksumString(address addr) internal pure returns (string memory) {
         // get the hash of the lowercase address
         bytes memory lowercaseAddr = ToLowercaseBytes(addr);
         bytes32 hashedAddr = keccak256(abi.encodePacked(lowercaseAddr));
@@ -25,8 +23,8 @@ library Address {
         uint160 hashValue = uint160(bytes20(hashedAddr));
 
         // checksum logic from EIP-55
-        for (uint i = 41; i > 1; --i) {
-            uint b = addrValue & 0xf;
+        for (uint256 i = 41; i > 1; --i) {
+            uint256 b = addrValue & 0xf;
 
             if (hashValue & 0xf > 7) {
                 result[i] = _CAPITAL[b];
@@ -42,13 +40,11 @@ library Address {
     }
 
     ///@dev get convert address bytes to lowercase char hex bytes (without '0x').
-    function ToLowercaseBytes(
-        address addr
-    ) internal pure returns (bytes memory) {
+    function ToLowercaseBytes(address addr) internal pure returns (bytes memory) {
         bytes memory s = new bytes(40);
 
         uint160 x = uint160(addr);
-        for (uint i = 0; i < 20; i++) {
+        for (uint256 i = 0; i < 20; i++) {
             uint8 b = uint8(x) & 0xff;
             s[39 - 2 * i - 1] = _LOWER[b >> 4]; // higher index
             s[39 - 2 * i] = _LOWER[b & 0xf]; // lower index
